[red-knot] refactor Definitions out of symbol table
